test: consolidate remaining coverage test files into primary test files
Merge 6 .coverage.test.* files into their existing counterparts:
- ScrollToTop.coverage.test.tsx → ScrollToTop.test.tsx
- CustomerPortalLink.native.coverage.test.tsx → CustomerPortalLink.native.test.tsx
- FeatureGate.web.coverage.test.tsx → FeatureGate.web.test.tsx
- UpgradePrompt.native.coverage.test.tsx → UpgradePrompt.native.test.tsx
- useUsage.coverage.test.tsx → useUsage.test.tsx
- openExternalUrl.native.coverage.test.ts → openExternalUrl.native.test.ts

ProtectedRoute.native.coverage.test.tsx is kept separate: its module-level
mock of AuthContext conflicts with the main test's AuthProvider setup.
