+++ apps/web/src/components/__tests__/ScrollToTop.test.tsx
@@ -141,6 +141,20 @@ describe('scrollToHashElement', () => {
   it('returns false for malformed percent-encoding without throwing', () => {
     expect(scrollToHashElement('#%E0%A4%A')).toBe(false);
   });
+
+  it('returns false when element is not in the DOM for a valid id', () => {
+    expect(scrollToHashElement('#not-in-dom-xyz')).toBe(false);
+  });
+
+  it('returns true and scrolls when element is found', () => {
+    const el = document.createElement('section');
+    el.id = 'coverage-anchor';
+    el.scrollIntoView = vi.fn();
+    document.body.append(el);
+    expect(scrollToHashElement('#coverage-anchor')).toBe(true);
+    expect(el.scrollIntoView).toHaveBeenCalled();
+    el.remove();
+  });
 });
 
 describe('startHashScroll', () => {
@@ -196,5 +210,24 @@ describe('startHashScroll', () => {
     }
     expect(target.scrollIntoView).not.toHaveBeenCalled();
     target.remove();
+  });
+
+  it('stops after HASH_SCROLL_MAX_FRAMES (60) without finding the element', () => {
+    const rafQueue: FrameRequestCallback[] = [];
+    vi.spyOn(window, 'requestAnimationFrame').mockImplementation(cb => {
+      rafQueue.push(cb);
+      return rafQueue.length;
+    });
+    vi.spyOn(console, 'warn').mockImplementation(() => {});
+
+    startHashScroll('#never-mounts-xyz');
+
+    for (let i = 0; i < 65; i++) {
+      const cb = rafQueue.shift();
+      if (!cb) break;
+      cb(i * 16);
+    }
+
+    expect(rafQueue).toHaveLength(0);
   });
 });
