--- packages/billing/src/components/CustomerPortalLink.native.coverage.test.tsx
@@ -1,48 +1,0 @@
-import { describe, expect, it, vi, beforeEach } from 'vitest';
-import { render, screen, fireEvent, waitFor } from '@testing-library/react';
-import React from 'react';
-import * as RN from 'react-native';
-import { CustomerPortalLink } from './CustomerPortalLink.native.js';
-import { useCustomerPortal } from '../hooks/useCustomerPortal.js';
-
-vi.mock('../hooks/useCustomerPortal.js', () => ({ useCustomerPortal: vi.fn() }));
-
-describe('CustomerPortalLink (native) — coverage', () => {
-  beforeEach(() => {
-    vi.spyOn(RN.Linking, 'openURL').mockResolvedValue(undefined as never);
-    vi.spyOn(RN.Linking, 'canOpenURL').mockResolvedValue(true as never);
-  });
-
-  it('renders React element children without an extra Text wrapper', () => {
-    vi.mocked(useCustomerPortal).mockReturnValue({ openPortal: vi.fn().mockResolvedValue(null), pending: false, error: null });
-    render(
-      <CustomerPortalLink>
-        <span data-testid='elem-child'>Manage</span>
-      </CustomerPortalLink>
-    );
-    expect(screen.getByTestId('elem-child')).toBeInTheDocument();
-  });
-
-  it('shows openError when openPortal throws and portalError is null', async () => {
-    vi.mocked(useCustomerPortal).mockReturnValue({
-      openPortal: vi.fn().mockRejectedValue(new Error('portal down')),
-      pending: false,
-      error: null,
-    });
-    render(<CustomerPortalLink>Portal</CustomerPortalLink>);
-    fireEvent.click(screen.getByText('Portal'));
-    await waitFor(() => expect(screen.getByText('portal down')).toBeInTheDocument());
-  });
-
-  it('suppresses openError when portalError is also set (openError && !portalError is false)', async () => {
-    vi.mocked(useCustomerPortal).mockReturnValue({
-      openPortal: vi.fn().mockRejectedValue(new Error('portal down')),
-      pending: false,
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      error: { kind: 'stripe', message: 'hook error' } as any,
-    });
-    render(<CustomerPortalLink>Portal</CustomerPortalLink>);
-    fireEvent.click(screen.getByText('Portal'));
-    await waitFor(() => expect(screen.queryByText('portal down')).not.toBeInTheDocument());
-  });
-});
