+++ packages/billing/src/components/UpgradePrompt.native.test.tsx
@@ -4,8 +4,12 @@ import React from 'react';
 import * as RN from 'react-native';
 import { UpgradePrompt } from './UpgradePrompt.native.js';
 import { useCheckout } from '../hooks/useCheckout.js';
+import { launchStripeCheckout } from '../utils/launchStripeCheckout.native.js';
 
 vi.mock('../hooks/useCheckout.js', () => ({ useCheckout: vi.fn() }));
+vi.mock('../utils/launchStripeCheckout.native.js', () => ({
+  launchStripeCheckout: vi.fn(),
+}));
 
 describe('UpgradePrompt (native)', () => {
   const startCheckout = vi.fn();
@@ -18,6 +22,7 @@ describe('UpgradePrompt (native)', () => {
       error: null,
     });
     vi.spyOn(RN.Linking, 'openURL').mockResolvedValue(undefined as never);
+    vi.mocked(launchStripeCheckout).mockResolvedValue(true);
   });
 
   it('opens checkout URL via Linking', async () => {
@@ -30,6 +35,40 @@ describe('UpgradePrompt (native)', () => {
       expect(RN.Linking.openURL).toHaveBeenCalledWith(
         'https://pay.example/start'
       );
+    });
+  });
+
+  it('renders via function children render-prop', () => {
+    render(
+      <UpgradePrompt targetTier='plan_pro' reason='Test'>
+        {({ onUpgrade, pending }) => (
+          <button onClick={() => void onUpgrade()} disabled={pending}>
+            custom-upgrade
+          </button>
+        )}
+      </UpgradePrompt>
+    );
+    expect(screen.getByText('custom-upgrade')).toBeInTheDocument();
+    expect(screen.queryByText('Upgrade')).not.toBeInTheDocument();
+  });
+
+  it('shows checkoutError.message as displayError when hook error is set', () => {
+    vi.mocked(useCheckout).mockReturnValue({
+      startCheckout,
+      pending: false,
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      error: { kind: 'stripe', message: 'hook-level-error' } as any,
+    });
+    render(<UpgradePrompt targetTier='plan_pro' reason='Upgrade reason' />);
+    expect(screen.getByText('hook-level-error')).toBeInTheDocument();
+  });
+
+  it('shows "Could not start checkout" when launchStripeCheckout returns false', async () => {
+    vi.mocked(launchStripeCheckout).mockResolvedValue(false);
+    render(<UpgradePrompt targetTier='plan_pro' reason='Upgrade reason' />);
+    fireEvent.click(screen.getByText('Upgrade'));
+    await waitFor(() => {
+      expect(screen.getByText('Could not start checkout')).toBeInTheDocument();
     });
   });
 });
